+++ apps/c/airfoil/airfoil_hdf5/qp/seq/bres_calc_seqkernel.cpp
@@ -51,11 +51,11 @@ void op_par_loop_bres_calc(char const *name, op_set set,
       map2idx = arg2.map_data[n * arg2.map->dim + 0];
 
       bres_calc(
-        &((double*)arg0.data)[2 * map0idx],
-        &((double*)arg0.data)[2 * map1idx],
-        &((double*)arg2.data)[4 * map2idx],
-        &((double*)arg3.data)[1 * map2idx],
-        &((double*)arg4.data)[4 * map2idx],
+        &((long double*)arg0.data)[2 * map0idx],
+        &((long double*)arg0.data)[2 * map1idx],
+        &((long double*)arg2.data)[4 * map2idx],
+        &((long double*)arg3.data)[1 * map2idx],
+        &((long double*)arg4.data)[4 * map2idx],
         &((int*)arg5.data)[1 * n]);
     }
   }
